feature: TS/33: code refactored for tasks
