+++ apps/admin/src/app/pages/tasks/list/index.js
@@ -1,0 +1,2 @@
+export * from './task-list.js';
+export * from './task-list.module.scss';
