+++ apps/admin/src/app/pages/tasks/list/task-list.js
@@ -26,7 +26,7 @@ import DownloadIcon from '@mui/icons-material/Download';
 import { styleNames } from 'libs/style-names';
 import { withErrorBoundary } from 'libs/error-boundary';
 
-import styles from './Tasks.module.scss';
+import styles from './task-list.module.scss';
 
 const sn = styleNames(styles);
